Loaded model - review

+++ db/migrate/20140504214331_create_reviews.rb
@@ -1,9 +1,9 @@
 class CreateReviews < ActiveRecord::Migration
   def change
     create_table :reviews do |t|
-      t.float :patty
-      t.float :bun
-      t.float :bun_patty_ratio
+      t.decimal :patty
+      t.decimal :bun
+      t.decimal :bun_patty_ratio
 
       t.timestamps
     end
